Concentration Measures

A set of different concentration measures.

comp(x, unbiased = FALSE, type = c("hhi", "hhi_d", "hhi_min", "dom",
 "sten", "all"),na.rm = TRUE)

hhi(x, unbiased = FALSE, na.rm = TRUE)

hhi_min(x, na.rm = TRUE)

hhi_d(x, unbiased = FALSE, na.rm = TRUE)

dom(x, unbiased = FALSE, na.rm = TRUE)

sten(x, na.rm = TRUE)

Arguments

x
A non-negative numeric vector.

unbiased
Logical. Argument specifying whether or not a finit sample correction should be applied.The default is FALSE.

type
A character string of the measure to be calculated, can be abbreviated with the first letter. Defaults to "hhi".

na.rm
A logical vector that indicates whether NA values should be excluded or not.

Value

prints the calculated measure

Details

comp is a wrapper for the proposed concentration measures hhi, hhi_d, hhi_min, dom, sten, all. If no measure is specified "hhi" will be the default.

hhi, hhi_min, hhi_d calculate the Herfindahl-Hirschman index, its minimum, and its dual, respectivly. dom calculates a dominance index and sten calculates the Stenbacka index. The index indicates the market share of a dominante position.

All measures can be accessed individually. hhi, hhi_d, and dom can be calculated individually as a normalized measure changing the default setting to TRUE. all computes all measures in a one step procedure.

Note

the vector of market shares should be in a decimal form corresponding to total share of individual firms/units.The sum of the vector should sum up to 1. You can also use sales figures to compute the respective measure.

References

Chang, E. J., Guerra, S. M., de Souza Peñaloza, R. A. & Tabak, B. M. (2005) Measuring Banking concentration: the Brazilian case. In Financial Stability Report. Brasilia: Banco Central do Brasil, 4: 109-129

García Alba Iduñate, P. (1994). Un índice de dominación para el análisis de la estructura de los mercados. El Trimestre Económico, 61: 499-524

Melnik, A., Shy, O. & Stenbacka, R. (2008). Assessing market dominance. Journal of Economic Behavior & Organization 68: 63-72

See also

Examples

# a vector of market shares
share <- c(0.35, 0.4, 0.05, 0.1, 0.06, 0.04)
# the Herfindahl-Hirschman index of the vector
share_hhi <- comp(share, type = "hhi")

#> [1] 0.3002

# individual measure
share_sten <- sten(share)

#> [1] 48.125

# complete group measures
share_comp <- comp(share, type = "all")

#>        Measure Value
#> 1          HHI  0.30
#> 2     HHI(min)  0.17
#> 3    HHI(dual)  0.44
#> 4    Dominance  0.45
#> 5 Stenbacka(%) 48.12