A set of Market Structure, Concentration, and Inequality Measures

a convenience function which calculates a selected set of different market structure, inequality and concentration measures more or less commonly used, e.g. k-firm ratios, Entropy, HHI, Palma ratio, and others in a one step procedure to provide a first overview.

concstats(x, na.rm = TRUE)

Arguments

x

a numeric vector.

na.rm

a logical vector that indicates whether NA values should be excluded or not. If set to FALSE the computation yields NA.

Value

returns a data frame of calculated measures .

Details

concstats computes a set of different and selected structural, inequality, and concentration measures in a one step procedure, however, all measures can be computed individually or in groups.

Note

the vector of market shares should be in a decimal form corresponding total share of individual firms/units.The vector should sum up to 1. You can also use sales figures to compute the respective measure.

See also

Examples

# a vector of market shares
# share <- c(0.35, 0.4, 0.05, 0.1, 0.06, 0.04)
# a selected set of different structural, concentration, and inequality
# measures
# share_conc <- concstats(share)