Inequality and Diversity Measures

A set of different inequality and diversity measures.

inequ(x, unbiased = FALSE, type = c("entropy", "gini", "berry", "palma",
 "grs", "all"), na.rm = TRUE)

entropy(x, unbiased = FALSE, na.rm = TRUE)

gini(x, unbiased = FALSE, na.rm = TRUE)

berry(x, unbiased = FALSE, na.rm = TRUE)

palma(x, na.rm = TRUE)

grs(x, na.rm = TRUE)

Arguments

x
a non-negative numeric vector.

unbiased
Logical. Argument specifying whether or not a finit sample correction should be applied.The default is FALSE.

type
a character string of the measure to be calculated, defaults to "entropy".

na.rm
a logical vector that indicates whether NA values should be excluded or not.

Value

prints the calculated measure

Details

inequ is a wrapper for the proposed inequality measures entropy, gini, berry, palma, grs, all. If no measure is specified, "entropy" is the default.

entropy returns the (Shannon) Entropy, gini is the Gini coefficient, berry is the inverse of the Herfindahl-Hirschman Index. You can normalize each of these three measures by setting unbiased = TRUE

Palma measures the ratio of inequality (used with income inequality) of the top 10 percent to the bottom 40 percent. grs is an alternative inequality measure (Ginevicius, 2009), all returns all measures in a one step procedure.

Note

the non- negative vector might be sales figures or market shares of individual firms/units. In the latter case the vector should sum up to 1.

References

C.E. Shannon (1948). A Mathematical Theory of Communication, The Bell System Technical Journal (Nokia Bell Labs)

J.G. Palma (2006). Globalizing Inequality: 'Centrifugal' and 'Centripetal' Forces at Work, DESA Working Paper No. 35.

A. Cobham / A. Summer (2013). Is It All About the Tails? The Palma Measure of Income Inequality, Center for Global Development, Washington, DC.

R. Ginevicius and S. Cirba (2009). Additive measurement of market concentration, Journal of Business Economics and Management, 10(3), 191-198. https://doi.org/10.3846/1611-1699.2009.10.191-198.

See also

Examples

# a vector of market shares
share <- c(0.4, 0.2, 0.25, 0.1, 0.05)
# Calculate the Palma ratio
share_p <- inequ(share, type = "palma")

#> [1] 2.666667

# Calculate the entropy measure directly
share_ent <- entropy(share, unbiased = TRUE)

#> [1] 0.879203

# Calculate the group measures
share_inequ <- inequ(share, type = "all")

#>       Measure Value
#> 1     Entropy  2.04
#> 2        Gini  0.34
#> 3 Berry Index  0.72
#> 4 Palma Ratio  2.67
#> 5         GRS  0.40