Market Structure Measures

Set of different measures to reflect a given market structure.

mstruct(x, type = c("firm", "nrs_eq", "top", "top3", "top5", "all"),
 na.rm = TRUE)

firm(x, na.rm = TRUE)

nrs_equ(x, na.rm = TRUE)

top(x, na.rm = TRUE)

top3(x, na.rm = TRUE)

top5(x, na.rm = TRUE)

Arguments

x
A non-negative numeric vector.

type
A character string of the measure to be calculated, can be abbreviated with the first letter. Defaults to "Firm".

na.rm
A logical vector that indicates whether NA values should be excluded or not.

Value

returns the calculated measure

Details

mstruct is a wrapper for the proposed structual measures firm, returns the number of firms with a given market share, nrs_eq computes the numbers equivalent, top, top3, and top5 calculate the share of the top (top 3 and top 5) firm(s). all computes all measures in a one step procedure. All measures can be computed individually.

Note

the vector of market shares should be in a decimal form corresponding to total share of individual firms/units.The sum of the vector should sum up to 1. You can also use sales figures to compute the respective measure.

See also

Examples

# a vector of market shares
share <- c(0.35, 0.4, 0.05, 0.1, 0.06, 0.04)
# the number of firms with market share
share_firm <- mstruct(share, type = "firm")

#> [1] 6

# Calculate top market share individually
share_top <- top(share)

#> [1] 40

# Calculate the market structure group measures
share_mstruct <- mstruct(share, type = "all")

#>          Measure value
#> 1          Firms  6.00
#> 2 Nrs_equivalent  3.33
#> 3        Top (%) 40.00
#> 4        Top3(%) 85.00
#> 5        Top5(%) 96.00