Inequality and Diversity Measures

A set of different inequality and diversity measures.

concstats_inequ(x, normalized = FALSE, type = c("entropy", "gini",
 "simpson", "palma", "grs", "all"), na.rm = TRUE)

concstats_entropy(x, normalized = TRUE, na.rm = TRUE)

concstats_gini(x, normalized = FALSE, na.rm = TRUE)

concstats_simpson(x, normalized = FALSE, na.rm = TRUE)

concstats_palma(x, na.rm = TRUE)

concstats_grs(x, na.rm = TRUE)

concstats_all_inequ(x, na.rm = TRUE)

Arguments

x

a non-negative numeric vector.

normalized

Logical. Argument specifying whether or not a finite sample correction should be applied. Must be either TRUE or FALSE. The default is FALSE.

type

a character string of the measure to be calculated, defaults to "concstats_entropy". Input is not case-sensitive.

na.rm

a logical vector that indicates whether NA values should be excluded or not. Must be either TRUE or FALSE. The default is TRUE. If set to FALSE the computation yields NA.

Value

the calculated numeric measure

a data.frame of inequality measures with default settings.

Details

concstats_inequ is a wrapper for the proposed inequality measures concstats_entropy, concstats_gini, concstats_simpson, concstats_palma, concstats_grs, concstats_all. If no measure is specified, "concstats_entropy" is the default.

concstats_entropy returns the Shannon Entropy (Shannon, 1948), concstats_gini is the Gini coefficient, concstats_simpson is the Gini-Simpson index (Simpson, 1949; Jost, 2006), also known as the Gini impurity (Gini's diversity index) in Machine Learning, Gibbs-Martin index or Blau index in sociology and management studies. You can normalize each of these three measures by setting normalized = TRUE concstats_palma measures the ratio of inequality (normally used with income inequality) of the top 10 percent to the bottom 40 percent (Palma, 2006). concstats_grs is an alternative inequality measure (Ginevicius, 2009) and concstats_all returns all measures in a one step procedure.

Note

the non-negative vector of market shares should be in a decimal form corresponding to the total shares of individual firms/units.The vector should sum up to 1.

References

Shannon, C. E. (1948). A Mathematical Theory of Communication, The Bell System Technical Journal (Nokia Bell Labs).

Simpson, E. H. (1949). Measurement of Diversity, Nature, 163, 688.

Palma, J. G. (2006). Globalizing Inequality: 'Centrifugal' and 'Centripetal' Forces at Work, DESA Working Paper No. 35.

Ginevicius, R. and S. Cirba (2009). Additive measurement of market concentration, Journal of Business Economics and Management, 10(3), 191-198.

Jost, L. (2006). Entropy and Diversity. Oikos, 113(2), 363-375.

See also

Examples

# a vector of market shares
x <- c(0.4, 0.2, 0.25, 0.1, 0.05)
# Calculate the Palma ratio
concstats_inequ(x, type = "palma")
#> [1] 2.666667
# Calculate the entropy measure directly
concstats_entropy(x, normalized = TRUE)
#> [1] 0.879203
# Calculate the group measures
concstats_inequ(x, type = "all")
#>         Measure Value
#> 1       Entropy  0.88
#> 2    Gini Index  0.34
#> 3 Simpson Index  0.72
#> 4   Palma Ratio  2.67
#> 5           GRS  0.40